processor/requirements.txt
opencv-python==4.10.0.82
pytesseract==0.3.10
layoutparser==0.3.4
boxdetect==1.0.2
pdf2image==1.17.0
pillow==10.3.0
numpy==1.26.4
click==8.1.7
rich==13.7.1
pytest==8.1.1
black==24.3.0
ruff==0.3.4 